Fix (?) broadcasting error
Note that this line hasn't changed in 3 years and I am only just now
getting an error after update. I suspect that something in the
denominator which used to be a scalar is now a vector, or vice versa.
Likely, there is a deeper bug hiding, this commit just prevents the
library from throwing.